Add documentation and restructure code
