+++ .gitignore
@@ -27,3 +27,5 @@ npm-debug.log.*
 *.css.d.ts
 *.sass.d.ts
 *.scss.d.ts
+
+/docs
